Almost completed this. Just a few more changes
